+++ src/sst/selftests/importing.py
@@ -1,4 +1,0 @@
-import _module
-
-# the current test directory should be added to sys.path
-assert _module.foo == 3